requirements.txt
loguru
rich_argparse
tabulate
expand_folders
scikit-learn
atlas-mpl-style
pytorch_lightning
matplotlib
seaborn
jupyterlab
numpy
pandas
tqdm
prettytable
dash
dash-bootstrap-components
python-multipart
dash-cytoscape
qiskit
qiskit-aer
pennylane
tensorflow
quimb
pylatexenc
pennylane-qiskit
kahypar
scipy
joblib
gdown
optuna
SQLAlchemy
xgboost
optuna-integration[xgboost]
geopandas
shapely
rasterio
pyarrow
pennylane-lightning
torchvision